Rename: pycore -> hdf5
h5group.py and h5dataset.py are now separate and work as a backend
module. Future changes should isolate the backend from the frontend, so
file.py would not import h5py
