+++ nixio/hdf5/h5group.py
